feat(server): convert recipe from dto to entity

+++ Server/Models/Food.cs
@@ -53,7 +53,7 @@ public class Measure
 {
   public int Id { get; set; }
   public MeasureType Type { get; set; }
-  public float Quantity { get; set; }
+  public double Quantity { get; set; }
 }
 
 public class Food
@@ -64,20 +64,20 @@ public class Food
   public string Description { get; set; } = string.Empty;
   public string DescriptionPt { get; set; } = string.Empty;
   public string Image { get; set; } = string.Empty;
-  public float Gi { get; set; }
-  public float Calories { get; set; }
-  public float Acidification { get; set; }
-  public float Carbohydrates { get; set; }
-  public float Ashes { get; set; }
-  public float Proteins { get; set; }
-  public float SaturedFats { get; set; }
-  public float MonounsaturatedFats { get; set; }
-  public float PolyunsaturatedFats { get; set; }
-  public float Cholesterol { get; set; }
-  public float TotalFat { get; set; }
-  public float DietaryFiber { get; set; }
-  public float Sugar { get; set; }
-  public float Gl { get; set; }
+  public double Gi { get; set; }
+  public double Calories { get; set; }
+  public double Acidification { get; set; }
+  public double Carbohydrates { get; set; }
+  public double Ashes { get; set; }
+  public double Proteins { get; set; }
+  public double SaturedFats { get; set; }
+  public double MonounsaturatedFats { get; set; }
+  public double PolyunsaturatedFats { get; set; }
+  public double Cholesterol { get; set; }
+  public double TotalFat { get; set; }
+  public double DietaryFiber { get; set; }
+  public double Sugar { get; set; }
+  public double Gl { get; set; }
   public UnitOfMeasurement UnitOfMeasurement { get; set; }
   public ICollection<Measure> OneMeasures { get; set; } = [];
 
@@ -87,62 +87,62 @@ public class Food
   public string Icon { get; set; } = string.Empty;
   public FoodType Type { get; set; } = FoodType.Solid;
   [NotMapped]
-  public float Quantity { get; set; } = 100;
+  public double Quantity { get; set; } = 100;
 
   // minerals
-  public float Calcium { get; set; }
-  public float Copper { get; set; }
-  public float Fluoride { get; set; }
-  public float Iron { get; set; }
-  public float Magnesium { get; set; }
-  public float Manganese { get; set; }
-  public float Phosphorus { get; set; }
-  public float Potassium { get; set; }
-  public float Sodium { get; set; }
-  public float Zinc { get; set; }
-  public float Selenium { get; set; }
+  public double Calcium { get; set; }
+  public double Copper { get; set; }
+  public double Fluoride { get; set; }
+  public double Iron { get; set; }
+  public double Magnesium { get; set; }
+  public double Manganese { get; set; }
+  public double Phosphorus { get; set; }
+  public double Potassium { get; set; }
+  public double Sodium { get; set; }
+  public double Zinc { get; set; }
+  public double Selenium { get; set; }
 
   // vitamins
-  public float A { get; set; }
-  public float AlphaCarotene { get; set; }
-  public float B1 { get; set; }
-  public float B11 { get; set; }
-  public float B12 { get; set; }
-  public float B2 { get; set; }
-  public float B3 { get; set; }
-  public float B5 { get; set; }
-  public float B6 { get; set; }
-  public float B7 { get; set; }
-  public float B9 { get; set; }
-  public float BetaCarotene { get; set; }
-  public float C { get; set; }
-  public float Choline { get; set; }
-  public float CryptoxanthinCarotene { get; set; }
-  public float D { get; set; }
-  public float D2 { get; set; }
-  public float D3 { get; set; }
-  public float E { get; set; }
-  public float K { get; set; }
-  public float Lycopene { get; set; }
+  public double A { get; set; }
+  public double AlphaCarotene { get; set; }
+  public double B1 { get; set; }
+  public double B11 { get; set; }
+  public double B12 { get; set; }
+  public double B2 { get; set; }
+  public double B3 { get; set; }
+  public double B5 { get; set; }
+  public double B6 { get; set; }
+  public double B7 { get; set; }
+  public double B9 { get; set; }
+  public double BetaCarotene { get; set; }
+  public double C { get; set; }
+  public double Choline { get; set; }
+  public double CryptoxanthinCarotene { get; set; }
+  public double D { get; set; }
+  public double D2 { get; set; }
+  public double D3 { get; set; }
+  public double E { get; set; }
+  public double K { get; set; }
+  public double Lycopene { get; set; }
 
   // amino acids
-  public float Alanine { get; set; }
-  public float Arginine { get; set; }
-  public float AsparticAcid { get; set; }
-  public float Cystine { get; set; }
-  public float GlutamicAcid { get; set; }
-  public float Glutamine { get; set; }
-  public float Glycine { get; set; }
-  public float Histidine { get; set; }
-  public float Isoleucine { get; set; }
-  public float Leucine { get; set; }
-  public float Lysine { get; set; }
-  public float Methionine { get; set; }
-  public float Phenylalanine { get; set; }
-  public float Proline { get; set; }
-  public float Serine { get; set; }
-  public float Threonine { get; set; }
-  public float Tryptophan { get; set; }
-  public float Tyrosine { get; set; }
-  public float Valine { get; set; }
+  public double Alanine { get; set; }
+  public double Arginine { get; set; }
+  public double AsparticAcid { get; set; }
+  public double Cystine { get; set; }
+  public double GlutamicAcid { get; set; }
+  public double Glutamine { get; set; }
+  public double Glycine { get; set; }
+  public double Histidine { get; set; }
+  public double Isoleucine { get; set; }
+  public double Leucine { get; set; }
+  public double Lysine { get; set; }
+  public double Methionine { get; set; }
+  public double Phenylalanine { get; set; }
+  public double Proline { get; set; }
+  public double Serine { get; set; }
+  public double Threonine { get; set; }
+  public double Tryptophan { get; set; }
+  public double Tyrosine { get; set; }
+  public double Valine { get; set; }
 }
